+++ src/hashtable.h
@@ -18,7 +18,7 @@ struct hashtable {
   struct fluffyvm* vm;
 
   int capacity;
-  int loadFactor;
+  double loadFactor;
   int usage;
   foxgc_object_t** table;
 
@@ -32,7 +32,7 @@ struct hashtable {
 bool hashtable_init(struct fluffyvm* vm);
 void hashtable_cleanup(struct fluffyvm* vm);
 
-struct hashtable* hashtable_new(struct fluffyvm* vm, int loadFactor, int initialCapacity, foxgc_root_t* root, foxgc_root_reference_t** rootRef);
+struct hashtable* hashtable_new(struct fluffyvm* vm, double loadFactor, int initialCapacity, foxgc_root_t* root, foxgc_root_reference_t** rootRef);
 // False if there an error
 bool hashtable_set(struct fluffyvm* vm, struct hashtable* this, struct value key, struct value value);
 struct value hashtable_get(struct fluffyvm* vm, struct hashtable* this, struct value key, foxgc_root_reference_t** rootRef);
